+++ core/src/main/scala/hammerlab/iterator/package.scala
@@ -2,6 +2,7 @@ package hammerlab
 
 package object iterator
   extends bulk
+     with count
      with scan
      with sliding
      with sorted
