+++ core/src/main/scala/org/hammerlab/iterator/count/CountByKey.scala
@@ -1,0 +1,20 @@
+package org.hammerlab.iterator.count
+
+import hammerlab.iterator.macros.IteratorWrapper
+
+import scala.collection.mutable
+
+@IteratorWrapper
+class CountByKey[K, V](it: Iterator[(K, V)]) {
+  def countByKey: Map[K, Int] = {
+    val counts = mutable.Map[K, Int]()
+    it.foreach {
+      case (k, _) ⇒
+        counts.update(
+          k,
+          counts.getOrElse(k, 0) + 1
+        )
+    }
+    counts.toMap
+  }
+}
